+++ packages/editor/js/extensions/libs/layout/components/gap.js
@@ -115,7 +115,7 @@ export default function ({
 						/>
 					)
 				) : (
-					<Grid gridTemplateColumns="55px 55px" gap="10px">
+					<Grid gridTemplateColumns="1fr 1fr" gap="8px">
 						<InputControl
 							columns="columns-1"
 							className="control-first label-center small-gap"