+++ packages/blocks-core/CHANGELOG.md
@@ -17,6 +17,7 @@
 - `List Item` block: `Icon` inner block added.
 - `Button` block: `Icon` inner block added.
 - `Page List` block: `Current Page` inner block added.
+- `Gallery` block: improve support.
 
 
 ### Automated Tests